backend/requirements.txt
fastapi
uvicorn
sentence-transformers
datasets
transformers
pydantic
pydantic-settings
sqlalchemy
alembic
psycopg2-binary
faiss-cpu
google-generativeai
